HTBHF-2028 Refactored PayloadMessageFactory methods into new component: EmailPayloadFactory
