+++ api/src/main/java/uk/gov/dhsc/htbhf/claimant/message/MessagePayloadFactory.java
@@ -2,15 +2,11 @@ package uk.gov.dhsc.htbhf.claimant.message;
 
 import uk.gov.dhsc.htbhf.claimant.entitlement.PaymentCycleVoucherEntitlement;
 import uk.gov.dhsc.htbhf.claimant.entity.Claim;
-import uk.gov.dhsc.htbhf.claimant.entity.Claimant;
 import uk.gov.dhsc.htbhf.claimant.entity.PaymentCycle;
 import uk.gov.dhsc.htbhf.claimant.message.payload.*;
 
 import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 import static uk.gov.dhsc.htbhf.claimant.message.MoneyUtils.convertPenceToPounds;
 
@@ -18,8 +14,6 @@ import static uk.gov.dhsc.htbhf.claimant.message.MoneyUtils.convertPenceToPounds
  * Factory object for building message payloads for emails.
  */
 public class MessagePayloadFactory {
-
-    private static final DateTimeFormatter DATE_FORMATTER = DateTimeFormatter.ofPattern("dd MMM yyyy");
 
     public static NewCardRequestMessagePayload buildNewCardMessagePayload(Claim claim,
                                                                           PaymentCycleVoucherEntitlement voucherEntitlement,
@@ -39,25 +33,6 @@ public class MessagePayloadFactory {
                 .build();
     }
 
-    /**
-     * Builds the message payload required to send an email message. The email template has parameterised values
-     * which are contained in the emailPersonalisation Map. All monetary amounts are formatted into pounds and the breakdown
-     * of voucher payments has been detailed in a bullet point list - any vouchers which are missing are replaced with a
-     * single blank line so that we don't have any empty bullet point in the email.
-     *
-     * @param paymentCycle The payment cycle with payment and voucher details.
-     * @param emailType The type of email to send.
-     * @return The constructed payload.
-     */
-    public static EmailMessagePayload buildEmailMessagePayload(PaymentCycle paymentCycle, EmailType emailType) {
-        Map<String, Object> emailPersonalisation = createPaymentEmailPersonalisationMap(paymentCycle, paymentCycle.getVoucherEntitlement());
-        return EmailMessagePayload.builder()
-                .claimId(paymentCycle.getClaim().getId())
-                .emailType(emailType)
-                .emailPersonalisation(emailPersonalisation)
-                .build();
-    }
-
     public static ReportClaimMessagePayload buildReportClaimMessagePayload(Claim claim) {
         return ReportClaimMessagePayload.builder()
                 .claimId(claim.getId())
@@ -71,44 +46,5 @@ public class MessagePayloadFactory {
         int totalAmount = numberOfVouchers * voucherAmountInPence;
         return String.format(summaryTemplate, convertPenceToPounds(totalAmount));
     }
-
-    private static Map<String, Object> createPaymentEmailPersonalisationMap(PaymentCycle paymentCycle, PaymentCycleVoucherEntitlement voucherEntitlement) {
-        Claimant claimant = paymentCycle.getClaim().getClaimant();
-        Map<String, Object> emailPersonalisation = new HashMap<>();
-        emailPersonalisation.put(EmailTemplateKey.FIRST_NAME.getTemplateKeyName(), claimant.getFirstName());
-        emailPersonalisation.put(EmailTemplateKey.LAST_NAME.getTemplateKeyName(), claimant.getLastName());
-        String paymentAmount = convertPenceToPounds(voucherEntitlement.getTotalVoucherValueInPence());
-        emailPersonalisation.put(EmailTemplateKey.PAYMENT_AMOUNT.getTemplateKeyName(), paymentAmount);
-        emailPersonalisation.put(EmailTemplateKey.PREGNANCY_PAYMENT.getTemplateKeyName(), buildPregnancyPaymentAmountSummary(voucherEntitlement));
-        emailPersonalisation.put(EmailTemplateKey.CHILDREN_UNDER_1_PAYMENT.getTemplateKeyName(), buildUnder1PaymentSummary(voucherEntitlement));
-        emailPersonalisation.put(EmailTemplateKey.CHILDREN_UNDER_4_PAYMENT.getTemplateKeyName(), buildUnder4PaymentSummary(voucherEntitlement));
-        String formattedCycleEndDate = paymentCycle.getCycleEndDate().plusDays(1).format(DATE_FORMATTER);
-        emailPersonalisation.put(EmailTemplateKey.NEXT_PAYMENT_DATE.getTemplateKeyName(), formattedCycleEndDate);
-        String backdateAmount = convertPenceToPounds(voucherEntitlement.getBackdatedVouchersValueInPence());
-        emailPersonalisation.put(EmailTemplateKey.BACKDATED_AMOUNT.getTemplateKeyName(), backdateAmount);
-        return emailPersonalisation;
-    }
-
-    private static String buildPregnancyPaymentAmountSummary(PaymentCycleVoucherEntitlement voucherEntitlement) {
-        return formatPaymentAmountSummary(
-                "\n* %s for a pregnancy",
-                voucherEntitlement.getVouchersForPregnancy(),
-                voucherEntitlement.getSingleVoucherValueInPence());
-    }
-
-    private static String buildUnder1PaymentSummary(PaymentCycleVoucherEntitlement voucherEntitlement) {
-        return formatPaymentAmountSummary(
-                "\n* %s for children under 1",
-                voucherEntitlement.getVouchersForChildrenUnderOne(),
-                voucherEntitlement.getSingleVoucherValueInPence());
-    }
-
-    private static String buildUnder4PaymentSummary(PaymentCycleVoucherEntitlement voucherEntitlement) {
-        return formatPaymentAmountSummary(
-                "\n* %s for children between 1 and 4",
-                voucherEntitlement.getVouchersForChildrenBetweenOneAndFour(),
-                voucherEntitlement.getSingleVoucherValueInPence());
-    }
-
 
 }
