+++ api/src/test/java/uk/gov/dhsc/htbhf/claimant/message/MessagePayloadFactoryTest.java
@@ -10,13 +10,8 @@ import java.time.LocalDate;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static uk.gov.dhsc.htbhf.claimant.message.EmailPayloadAssertions.assertEmailPayloadCorrectForClaimantWithAllVouchers;
-import static uk.gov.dhsc.htbhf.claimant.message.EmailPayloadAssertions.assertEmailPayloadCorrectForClaimantWithPregnancyVouchersOnly;
-import static uk.gov.dhsc.htbhf.claimant.message.MessagePayloadFactory.buildEmailMessagePayload;
 import static uk.gov.dhsc.htbhf.claimant.message.MessagePayloadFactory.buildReportClaimMessagePayload;
 import static uk.gov.dhsc.htbhf.claimant.testsupport.ClaimTestDataFactory.aValidClaim;
-import static uk.gov.dhsc.htbhf.claimant.testsupport.PaymentCycleTestDataFactory.aPaymentCycleWithPregnancyVouchersOnly;
-import static uk.gov.dhsc.htbhf.claimant.testsupport.PaymentCycleTestDataFactory.aPaymentCycleWithStartAndEndDate;
 import static uk.gov.dhsc.htbhf.claimant.testsupport.PaymentCycleTestDataFactory.aValidPaymentCycle;
 import static uk.gov.dhsc.htbhf.claimant.testsupport.PaymentCycleVoucherEntitlementTestDataFactory.aPaymentCycleVoucherEntitlementWithVouchers;
 
@@ -43,60 +38,6 @@ class MessagePayloadFactoryTest {
         assertThat(payload.getClaimId()).isEqualTo(paymentCycle.getClaim().getId());
         assertThat(payload.getPaymentCycleId()).isEqualTo(paymentCycle.getId());
         assertThat(payload.getCardAccountId()).isEqualTo(paymentCycle.getClaim().getCardAccountId());
-    }
-
-    @Test
-    void shouldBuildSendNewCardSuccessEmailPayloadWithAllPaymentTypes() {
-
-        LocalDate startDate = LocalDate.now();
-        LocalDate endDate = startDate.plusDays(28);
-        PaymentCycle paymentCycle = aPaymentCycleWithStartAndEndDate(startDate, endDate);
-
-        EmailMessagePayload payload = buildEmailMessagePayload(paymentCycle, EmailType.NEW_CARD);
-
-        assertThat(payload.getClaimId()).isEqualTo(paymentCycle.getClaim().getId());
-        assertThat(payload.getEmailType()).isEqualTo(EmailType.NEW_CARD);
-        assertEmailPayloadCorrectForClaimantWithAllVouchers(payload.getEmailPersonalisation(), endDate.plusDays(1));
-    }
-
-    @Test
-    void shouldBuildSendNewCardSuccessEmailPayloadWithOnlyPregnancyPayment() {
-        LocalDate startDate = LocalDate.now();
-        LocalDate endDate = startDate.plusDays(28);
-        PaymentCycle paymentCycle = aPaymentCycleWithPregnancyVouchersOnly(startDate, endDate);
-
-        EmailMessagePayload payload = buildEmailMessagePayload(paymentCycle, EmailType.NEW_CARD);
-
-        assertThat(payload.getClaimId()).isEqualTo(paymentCycle.getClaim().getId());
-        assertThat(payload.getEmailType()).isEqualTo(EmailType.NEW_CARD);
-        assertEmailPayloadCorrectForClaimantWithPregnancyVouchersOnly(payload.getEmailPersonalisation(), endDate.plusDays(1));
-    }
-
-    @Test
-    void shouldBuildPaymentNotificationEmailPayloadWithAllPaymentTypes() {
-
-        LocalDate startDate = LocalDate.now();
-        LocalDate endDate = startDate.plusDays(28);
-        PaymentCycle paymentCycle = aPaymentCycleWithStartAndEndDate(startDate, endDate);
-
-        EmailMessagePayload payload = buildEmailMessagePayload(paymentCycle, EmailType.PAYMENT);
-
-        assertThat(payload.getClaimId()).isEqualTo(paymentCycle.getClaim().getId());
-        assertThat(payload.getEmailType()).isEqualTo(EmailType.PAYMENT);
-        assertEmailPayloadCorrectForClaimantWithAllVouchers(payload.getEmailPersonalisation(), endDate.plusDays(1));
-    }
-
-    @Test
-    void shouldBuildPaymentNotificationEmailPayloadWithOnlyPregnancyPayment() {
-        LocalDate startDate = LocalDate.now();
-        LocalDate endDate = startDate.plusDays(28);
-        PaymentCycle paymentCycle = aPaymentCycleWithPregnancyVouchersOnly(startDate, endDate);
-
-        EmailMessagePayload payload = buildEmailMessagePayload(paymentCycle, EmailType.PAYMENT);
-
-        assertThat(payload.getClaimId()).isEqualTo(paymentCycle.getClaim().getId());
-        assertThat(payload.getEmailType()).isEqualTo(EmailType.PAYMENT);
-        assertEmailPayloadCorrectForClaimantWithPregnancyVouchersOnly(payload.getEmailPersonalisation(), endDate.plusDays(1));
     }
 
     @Test
